Update BuildStream elements and scripts to support a MacOS build

+++ src/libpvparallelview/include/pvparallelview/PVQuadTree.h
@@ -1081,7 +1081,7 @@ class PVQuadTree
 			const uint64_t max_count = 1 << zoom;
 			const uint64_t y1_orig = obj._y1_min_value;
 			const uint64_t y1_len = (obj._y1_mid_value - y1_orig) * 2;
-			const uint64_t y1_scale = std::max(1UL, y1_len / max_count);
+			const uint64_t y1_scale = std::max((uint64_t)1, y1_len / max_count);
 
 			const uint64_t y2_orig = obj._y2_min_value;
 			const uint64_t y2_scale = ((obj._y2_mid_value - y2_orig) * 2) / y2_count;
@@ -1090,7 +1090,7 @@ class PVQuadTree
 			    (PVCore::clamp(y1_min, y1_orig, y1_orig + y1_len) - y1_orig) / y1_scale;
 			const uint64_t ly1_max =
 			    (PVCore::clamp(y1_max, y1_orig, y1_orig + y1_len) - y1_orig) / y1_scale;
-			const uint64_t clipped_max_count = std::max(1UL, ly1_max - ly1_min);
+			const uint64_t clipped_max_count = std::max((uint64_t)1, ly1_max - ly1_min);
 			const size_t count_aligned = ((clipped_max_count * y2_count) + 31) / 32;
 			memset(buffer, 0, count_aligned * sizeof(uint32_t));
 			uint32_t remaining = clipped_max_count * y2_count;
@@ -1158,7 +1158,7 @@ class PVQuadTree
 			const uint64_t max_count = 1 << zoom;
 			const uint64_t y1_orig = obj._y1_min_value;
 			const uint64_t y1_len = (obj._y1_mid_value - y1_orig) * 2;
-			const uint64_t y1_scale = std::max(1UL, y1_len / max_count);
+			const uint64_t y1_scale = std::max((uint64_t)1, y1_len / max_count);
 			const uint64_t y1_shift = log2(y1_scale);
 			const uint64_t y2_orig = obj._y2_min_value;
 			const uint64_t y2_scale = ((obj._y2_mid_value - y2_orig) * 2) / y2_count;
@@ -1167,7 +1167,7 @@ class PVQuadTree
 			    (PVCore::clamp(y1_min, y1_orig, y1_orig + y1_len) - y1_orig) / y1_scale;
 			const uint64_t ly1_max =
 			    (PVCore::clamp(y1_max, y1_orig, y1_orig + y1_len) - y1_orig) / y1_scale;
-			const uint64_t clipped_max_count = std::max(1UL, ly1_max - ly1_min);
+			const uint64_t clipped_max_count = std::max((uint64_t)1, ly1_max - ly1_min);
 			const size_t count_aligned = ((clipped_max_count * y2_count) + 31) / 32;
 			memset(buffer, 0, count_aligned * sizeof(uint32_t));
 			uint32_t remaining = clipped_max_count * y2_count;
@@ -1535,12 +1535,12 @@ class PVQuadTree
 			const uint64_t y1_scale = ((obj._y1_mid_value - y1_orig) * 2) / y1_count;
 			const uint64_t y2_orig = obj._y2_min_value;
 			const uint64_t y2_len = (obj._y2_mid_value - y2_orig) * 2;
-			const uint64_t y2_scale = std::max(1UL, y2_len / max_count);
+			const uint64_t y2_scale = std::max((uint64_t)1, y2_len / max_count);
 			const uint64_t ly2_min =
 			    (PVCore::clamp(y2_min, y2_orig, y2_orig + y2_len) - y2_orig) / y2_scale;
 			const uint64_t ly2_max =
 			    (PVCore::clamp(y2_max, y2_orig, y2_orig + y2_len) - y2_orig) / y2_scale;
-			const uint64_t clipped_max_count = std::max(1UL, ly2_max - ly2_min);
+			const uint64_t clipped_max_count = std::max((uint64_t)1, ly2_max - ly2_min);
 			const size_t count_aligned = ((clipped_max_count * y1_count) + 31) / 32;
 			memset(buffer, 0, count_aligned * sizeof(uint32_t));
 			uint32_t remaining = clipped_max_count * y1_count;
@@ -1596,13 +1596,13 @@ class PVQuadTree
 			const uint64_t y1_shift = log2(y1_scale);
 			const uint64_t y2_orig = obj._y2_min_value;
 			const uint64_t y2_len = (obj._y2_mid_value - y2_orig) * 2;
-			const uint64_t y2_scale = std::max(1UL, y2_len / max_count);
+			const uint64_t y2_scale = std::max((uint64_t)1, y2_len / max_count);
 			const uint64_t y2_shift = log2(y2_scale);
 			const uint64_t ly2_min =
 			    (PVCore::clamp(y2_min, y2_orig, y2_orig + y2_len) - y2_orig) / y2_scale;
 			const uint64_t ly2_max =
 			    (PVCore::clamp(y2_max, y2_orig, y2_orig + y2_len) - y2_orig) / y2_scale;
-			const uint64_t clipped_max_count = std::max(1UL, ly2_max - ly2_min);
+			const uint64_t clipped_max_count = std::max((uint64_t)1, ly2_max - ly2_min);
 			const size_t count_aligned = ((clipped_max_count * y1_count) + 31) / 32;
 			memset(buffer, 0, count_aligned * sizeof(uint32_t));
 			uint32_t remaining = clipped_max_count * y1_count;
